refactor: 레이어드 아키텍쳐 적용
