+++ src/main/java/roomescape/dto/ReservationTimeCreateRequestDto.java
@@ -1,7 +1,7 @@
 package roomescape.dto;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
-import roomescape.model.ReservationTime;
+import roomescape.domain.ReservationTime;
 
 import java.time.LocalTime;
