+++ src/test/java/roomescape/domain/ReservationTest.java
@@ -1,4 +1,4 @@
-package roomescape.model;
+package roomescape.domain;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
